### 2025-08-04 #8711 @editorialbot
[PRE REVIEW]: `INS_IEP`: A MATLAB package for fitting Inelastic Neutron Scattering data
**Submitting author:** @AlbanBloorRiley (<a href="http://orcid.org/0000-0001-9431-709X">Alban Bloor Riley</a>)
**Repository:** https://github.com/AlbanBloorRiley/INS_IEP
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
