### 2025-08-04 c22c434
Update Examples.m
