### 2025-08-04 efc8e13
Update paper.bib
