+++ Paper/paper.bib
@@ -126,6 +126,7 @@
 	year = {2020},
 	note = {Publisher: Taylor \& Francis},
 	pages = {1248--1271},
+	doi = {10.1080/10556788.2019.1613655},
 } 
 
 
