### 2025-08-04 #8711 @AlbanBloorRiley
Hi there @arfon,
Thanks for the update, I understand that this paper is currently tagged as in the Biomedical Engineering, Biosciences, Chemistry, and Materials area of expertise, but it could also be within the Mathematics section if that is easier?
No worries if not, thank you for your help.
Cheers
