### 2025-10-03 #8711 @mbarzegary
Hi @stuartcampbell, @mdoucet, @cmbiwer, @andyfaff, @KedoKudo 👋
Considering your expertise, would any of you be willing to review this submission for JOSS? As you may know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.08711/joss.08711/10.21105.joss.08711.pdf
Software: https://github.com/AlbanBloorRiley/INS_IEP
