### 2025-10-05 #8711 @andyfaff
@mbarzegary, sorry I also have no familiarity with Matlab. Perhaps @granrothge or @RIchardWaiteSTFC (contributors to [SpinW](https://github.com/SpinW/spinw)) might be able to review.
