### 2025-10-31 AlbanBloorRiley/INS_IEP#2 @granrothge
Improve the installation instructions.
I find the installation instructions a little to tearse.

* I would suggest explicitly adding that you need to install easyspin and move it as step 1.
* I would specify which directories you need to add to the matlab path.

openjournals/joss-reviews/issues/9176
