### 2025-10-31 AlbanBloorRiley/INS_IEP#3 @granrothge
iitest ends in a crash
Running iitest after installation  produces the following crash
openjournals/joss-reviews/issues/9176

```>> iitest
=======================================================================
INS_IEP test set                      31-Oct-2025 14:45:48
(MATLAB 24.2.0.2923080 (R2024b) Update 6)
INS_IEP location: /home/19g/repos/INS_IEP
=======================================================================
INS_IEP                                          direct      crashed  1
    Unrecognized function or variable 'Opt'.

    Error in iitest (line 116)
            ok = testFcn(Opt);
                         ^^^
-----------------------------------------------------------------------
INS_IEP                                          direct      crashed  1
    Unrecognized function or variable 'Opt'.

    Error in iitest (line 116)
            ok = testFcn(Opt);
                         ^^^
-----------------------------------------------------------------------
0 passes, 0 failures, 1 crashes
-----------------------------------------------------------------------```
