### 2025-11-03 AlbanBloorRiley/INS_IEP#4 @granrothge
Improvements to Mn12 Example in paper
Here are a few things related to the Mn 12 Example.

- [ ]  I believe the units are GHz not Hz.  This would make sense with Easy Spin being an EPR package.
- [ ]  It would help to reference where the experimental Eigenvalues came from here. Maybe the Bircher et al. reference?
- [ ] In the Bircher et al. paper (and the previous studies) Only the 1.24meV mode is identified.  Where did the other modes come from?

openjournals/joss-reviews/issues/9176
