+++ Paper/paper.bib
@@ -176,20 +176,22 @@
 }
 
 
-
-@misc{bloor_riley_deflation_2025,
+@article{bloor_riley_deflation_2025,
 	title = {Deflation {Techniques} for {Finding} {Multiple} {Local} {Minima} of a {Nonlinear} {Least} {Squares} {Problem}},
-	url = {http://arxiv.org/abs/2409.14438},
-	doi = {10.48550/arXiv.2409.14438},
-	abstract = {In this paper we generalize the technique of deflation to define two new methods to systematically find many local minima of a nonlinear least squares problem. The methods are based on the Gauss-Newton algorithm, and as such do not require the calculation of a Hessian matrix. They also require fewer deflations than for applying the deflated Newton method on the first order optimality conditions, as the latter finds all stationary points, not just local minima. One application of interest covered in this paper is the inverse eigenvalue problem (IEP) associated with the modelling of spectroscopic data of relevance to the physical and chemical sciences. Open source MATLAB code is provided at https://github.com/AlbanBloorRiley/DeflatedGaussNewton.},
-	urldate = {2025-06-17},
-	publisher = {arXiv},
+	volume = {47},
+	issn = {1064-8275, 1095-7197},
+	url = {https://epubs.siam.org/doi/10.1137/24M1695191},
+	doi = {10.1137/24M1695191},
+	language = {en},
+	number = {5},
+	urldate = {2025-11-05},
+	journal = {SIAM Journal on Scientific Computing},
 	author = {Bloor Riley, Alban and Webb, Marcus and Baker, Michael L.},
-	month = jun,
+	month = oct,
 	year = {2025},
-	note = {arXiv:2409.14438 [math]},
-	file = {Preprint PDF:/Users/user/Zotero/storage/KLYXA9D7/Riley et al. - 2025 - Deflation Techniques for Finding Multiple Local Mi.pdf:application/pdf},
+	pages = {A2963--A2987},
 }
+
 
 @misc{bloor_riley_riemannian_2025,
 	title = {A {Riemannian} {Gradient} {Descent} {Method} for the {Least} {Squares} {Inverse} {Eigenvalue} {Problem}},
