### 2025-11-05 AlbanBloorRiley/INS_IEP#4 @AlbanBloorRiley
Yes you are right, the unit is GHz, I've changed it.
The eigenvalues are actually just simulated from the parameters given in the paper - I have added a note to explain this.
But I think all the modes are given in Fig 3?
