### 2025-11-05 AlbanBloorRiley/INS_IEP#4 @granrothge
Ok the third bullet is fine.  It was a misunderstanding in what your code is trying to do.  I will add another issue regarding paper improvements that will cover this.
