### 2025-11-05 AlbanBloorRiley/INS_IEP#5 @granrothge
Paper considerations
- The title needs to be updated.
    - [ ] First to reflect that the code does not fit INS data directly, but starts from eigenvalues derived from the INS data.
    - [ ] Second to reflect that you are dealing with molecular magnets or spin clusters.

-  [ ] There should be a direct reference to the state of the art iterative method.
-  [ ] The authors should consider if a reference to exact diagonalisation methods is warranted.

openjournals/joss-reviews/issues/9176
