### 2025-11-05 AlbanBloorRiley/INS_IEP#6 @granrothge
Add Community guidelines.
The authors have done a nice job of including a code of conduct for contributors.
However I haven't been able to find any guidelines for people wishing to contirbute to the code, or for users needing help, or for bug reporting.
openjournals/joss-reviews/issues/9176
