### 2025-11-23 AlbanBloorRiley/INS_IEP#7 @KedoKudo
Installation instructions cause test failure
Following the installation instructions in README.md (line 8) causes the test suite to fail. The instructions specify adding `INS_IEP/INS_IEP` to the MATLAB path, but the main function `INS_IEP.m` is located in the root directory, not the subdirectory.

**Reproduction:**
```matlab
% Following README.md line 8
addpath('/path/to/INS_IEP/INS_IEP')
cd Tests
iitest
```

**Result:**
```
error in using iitest (line 27)
INS_IEP is not on the MATLAB path
```

**Root cause:** Repository structure has `INS_IEP.m` in root, but instructions only add the subdirectory to path. The test file checks for `INS_IEP.m` using `which('INS_IEP')` and fails.

**Checklist:**
- [x] Update README.md to specify adding the root directory to MATLAB path
- [x] Provide explicit MATLAB commands (e.g., `addpath(genpath('/path/to/INS_IEP'))`)
- [x] Add verification step showing how to check installation
- [ ] Consider adding a `setup.m` script to automate path configuration
- [x] Test updated instructions in fresh MATLAB session

**Related:** #2

Reported in JOSS review openjournals/joss-reviews#9176
