### 2025-11-23 AlbanBloorRiley/INS_IEP#8 @KedoKudo
Documentation corrections
Multiple minor documentation issues identified during review:

**Typos:**
- [ ] Line 6 @ README.md: Dowload → Download
- [ ] Line 15 @ README.md: MTLAB → MATLAB

**Missing information:**
- [ ] Line 63 @ .github/CODE_OF_CONDUCT.md: Add enforcement contact email address

**PDF rendering:**
- [ ] Paper/paper.pdf footer shows unicode errors: `¿VOL?(¿ISSUE?), ¿PAGE?`
  - Investigate if this is reproducible with current JOSS template
  - May be a JOSS template issue rather than author error

Reported in JOSS review openjournals/joss-reviews#9176
