### 2025-11-23 AlbanBloorRiley/INS_IEP#9 @KedoKudo
Paper missing comparison to existing approaches
JOSS review criteria require papers to describe how the software compares to other commonly-used packages or approaches. The paper describes the current manual fitting approach and states this is the first IEP formulation for INS data, but lacks explicit comparison.

**Current state:**
- Manual fitting approach is described qualitatively
- No quantitative performance comparison provided
- No discussion of why general-purpose optimization tools are insufficient

**Checklist:**
- [x] Add section explicitly comparing to current practice
- [ ] Provide quantitative example (e.g., for Example 1: manual approach requires ~X trials/hours, INS_IEP finds all 4 solutions in Y minutes)
- [x] Clarify that no other systematic INS fitting software exists (if true)
- [ ] Briefly explain why general-purpose tools (fmincon, scipy.optimize) are insufficient for this problem
- [x] Consider adding reference to state-of-art iterative methods (relates to #5)
- [x] Address exact diagonalization methods if relevant (relates to #5)

**Related:** #5

Reported in JOSS review openjournals/joss-reviews#9176
