### 2025-11-23 AlbanBloorRiley/INS_IEP#10 @KedoKudo
Remove .DS_Store from version control
The repository tracks a `.DS_Store` file (macOS system file), which should not be version controlled.

The file is already in `.gitignore` (line 9) but was committed before being ignored.

Please do the following:

**Resolution:**
```bash
git rm --cached .DS_Store
git commit -m "Remove .DS_Store from version control"
```

The file will remain in `.gitignore` and won't be tracked again.

Reported in JOSS review openjournals/joss-reviews#9176
