### 2025-11-23 #9176 @KedoKudo
@AlbanBloorRiley
Thank you for submitting INS_IEP. This is strong work addressing a genuine need in the INS community with novel algorithmic contributions. The software is well-implemented with good test coverage and clear examples. Here are some new issues I created to help further improve it:

- https://github.com/AlbanBloorRiley/INS_IEP/issues/7 (critical)
- https://github.com/AlbanBloorRiley/INS_IEP/issues/8 (minor)
- https://github.com/AlbanBloorRiley/INS_IEP/issues/9 (high priority)
- https://github.com/AlbanBloorRiley/INS_IEP/issues/10 (minor)

> I've reviewed existing issues by @granrothge and linked to them where appropriate

Feel free to pin me if you have any questions about these points. The software is in excellent shape and these are primarily documentation improvements.
