### 2025-11-24 b54936c
Add contribution and support guidelines to README
Added sections to the README detailing how to contribute, report issues, and seek support. This provides clearer instructions for users and potential contributors.
