+++ README.md
@@ -23,3 +23,16 @@ INS_IEP/      source code for the package
 Examples/        collection of INS_IEP example scripts
 Paper/        source code for the JOSS paper (under review)
 Tests/        unit tests for INS_IEP
+
+
+## How to Contribute to this Software
+
+To make changes to the code, please make a fork of the repository, push your changes there and then make a pull request in this respository. Your contribution will be reviewed by one of the current contributors of the repository before being merged with the repository.
+
+## How to Report Issues, Bugs or Feature Requests
+
+Issues, bugs and feature requests should be reported using the GitHub Issues functionality.
+
+## Seeking Support
+
+If, after consulting the example scripts in the Examples folder, you need assistance using this software then please post an Issue using the GitHub Issues functionality, and we will endeavor to assist and/or improve the documentation of the code.
