### 2025-11-24 dbb22bd
Updated installation instructions
