### 2025-11-24 a745326
Clarify we solve IEP vs full fitting
Added explanation that we solve the inverse eigenvalue problem (IEP) rather than fitting the full INS data, as an initial proxy.
