+++ Paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-Inelastic neutron scattering (INS) is a spectroscopic technique that can be used to measure the magnetic excitations in materials with interacting electron spins. For samples composed of finite-size clusters of magnetic moment-carrying atoms, such as single ions or molecular-based magnets, INS experiments yield inelastic excitation with energies that correspond to the eigenvalues of the spin Hamiltonian of the material being studied. Fitting a model spin system to these experimental eigenvalues can be formulated as an inverse eigenvalue problem (IEP), where the matrix formed is the spin Hamiltonian operator of the sample molecule. `INS_IEP` is a MATLAB package that uses deflated numerical optimisation methods to systematically find multiple solutions to this IEP. The package requires and is fully compatible with `easyspin` [@stoll_easyspin_2006], a package for solving fitting problems in electron paramagnetic resonance (EPR).
+Inelastic neutron scattering (INS) is a spectroscopic technique that can be used to measure the magnetic excitations in materials with interacting electron spins. For samples composed of finite-size clusters of magnetic moment-carrying atoms, such as single ions or molecular-based magnets, INS experiments yield inelastic excitation with energies that correspond to the eigenvalues of the spin Hamiltonian of the material being studied. Fitting a model spin system to these experimental eigenvalues can be formulated as an inverse eigenvalue problem (IEP), where the matrix formed is the spin Hamiltonian operator of the sample molecule. Solving this IEP is less computationally expensive than fitting the full INS data, so can be used as an initial proxy for the full fitting problem. `INS_IEP` is a MATLAB package that uses deflated numerical optimisation methods to systematically find multiple solutions to this IEP. The package requires and is fully compatible with `easyspin` [@stoll_easyspin_2006], a package for solving fitting problems in electron paramagnetic resonance (EPR).
 
 # Statement of need
 
