### 2025-11-24 AlbanBloorRiley/INS_IEP#2 @AlbanBloorRiley
I have changed the installation instructions to make them more clear.
