### 2025-11-24 362d45e
Added exact diagonalisation methods to the Methods section.
