+++ Paper/paper.md
@@ -106,6 +106,8 @@ All of the methods used are iterative schemes of the form $x^{k+1} = x^k +p^k$ w
 
 Where the matrix $B$ is the Gram matrix formed from the frobenius inner products of the basis matrices: $B_{ij} = \langle A_i, A_j\rangle_F$. The Lift and Projection method is a Riemannian Gradient descent method [@bloor_riley_riemannian_2025], inspired by the Lift and Projection method [@chu_inverse_2005], specifically designed for solving IEPs. In [@bloor_riley_riemannian_2025] it is proven that the method is a strictly descending algorithm, that is it reduces the value of the objective function every step. Both the deflated Gauss-Newton method and the Riemannian Gradient descent Lift and Projection method are new methods designed for this package [@bloor_riley_deflation_2025;@bloor_riley_riemannian_2025]. 
 
+The $m$ eigenvalues of required to calculate $J_r$ are calculated using MATLAB's `eigs'' command, which is an efficient method based on Krylov subspaces to calculate a small subset of the eigenvalues of the matrix. This is invaluable in this setting because computing the full set of eigenvalues by exact diagonalisation is in some cases infeasible [@bloor_riley_riemannian_2025].
+
 ## Deflation 
 
 The number of eigenvalues that can be probed via INS experiments varies  depending on the equipment and sample in question, meaning that the fitting problem is often under or even over  determined. The IEP is also highly nonlinear and due to the experimental nature of the data may be ill-posed. One consequence of this is that the solution space may be very 'bumpy', that is there may exist many local minimisers to the problem. For example in \autoref{fig:contour}, there are clearly 4 distinct solutions (for more details see Example 1 and the file Example1_Mn12.m in the examples folder). We seek to solve the problem of multiple local minima by the use of Deflation, a numerical technique used to find multiple solutions to systems of equations [@farrell_deflation_2015]. Fortunately it is cheap to apply deflation for the above methods, it is simply a change to the length of the step - notably this means that the direction of each step does not change. It is  proven in [@bloor_riley_deflation_2025] that the deflated methods will not converge to deflated points.  The usual requirements still apply to the convergence of the new methods - that the initial guess is close enough to the new minimum, and that the Jacobian is full rank in a neighbourhood around that minimum. The rate of convergence of the deflated methods is also more complicated, although the number of iterations required to converge can go up with the number of deflations this is not a strict correlation, as can be seen in  \autoref{fig:convergence}.
