### 2025-11-24 bad4222
Corrected eigs typo
