### 2025-11-24 48f01a6
Corrected Newton's method typo
