### 2025-11-29 38b7ce5
Update paper.bib
