+++ Paper/paper.bib
@@ -329,3 +329,44 @@
 	author = {Baker, Michael L.},
 	year = {2022},
 }
+
+
+
+@misc{a_t_boothroyd_spectre_1990,
+	title = {Spectre - {A} program for calculating spectroscopic prop- erties of rare earth ions in crystals},
+	author = {{A. T. Boothroyd}},
+	year = {1990},
+}
+
+@article{fabi_focus_1995,
+	title = {{FOCUS}: {An} {Interactive} crystal electric field parameter fitting package using neutron scattering data},
+	author = {Fabi, Peter},
+	month = jun,
+	year = {1995},
+}
+
+@article{scheie_pycrystalfield_2021,
+	title = {{PyCrystalField} - software for calculation, analysis and fitting of crystal electric field {Hamiltonians}},
+	volume = {54},
+	issn = {1600-5767},
+	shorttitle = {\textit{{PyCrystalField}}},
+	url = {https://journals.iucr.org/paper?S160057672001554X},
+	doi = {10.1107/S160057672001554X},
+	abstract = {PyCrystalField
+              is a Python software package for calculating single-ion crystal electric field (CEF) Hamiltonians. This software can calculate a CEF Hamiltonian
+              ab initio
+              from a point charge model for any transition or rare earth ion in either the
+              J
+              basis or the
+              LS
+              basis, perform symmetry analysis to identify nonzero CEF parameters, calculate the energy spectrum and observables such as neutron spectrum and magnetization, and fit CEF Hamiltonians to any experimental data. The theory, implementation and examples of its use are discussed.},
+	number = {1},
+	journal = {Journal of Applied Crystallography},
+	author = {Scheie, A.},
+	month = feb,
+	year = {2021},
+	pages = {356--362},
+}
+
+
+
