+++ Paper/paper.bib
@@ -332,18 +332,22 @@
 
 
 
+
 @misc{a_t_boothroyd_spectre_1990,
 	title = {Spectre - {A} program for calculating spectroscopic prop- erties of rare earth ions in crystals},
+	url = {https://www.physics.ox.ac.uk/our-people/boothroyd/software},
 	author = {{A. T. Boothroyd}},
 	year = {1990},
 }
 
 @article{fabi_focus_1995,
 	title = {{FOCUS}: {An} {Interactive} crystal electric field parameter fitting package using neutron scattering data},
+	url = {https://www.isis.stfc.ac.uk/Pages/focus-manual6828.pdf},
 	author = {Fabi, Peter},
 	month = jun,
 	year = {1995},
 }
+
 
 @article{scheie_pycrystalfield_2021,
 	title = {{PyCrystalField} - software for calculation, analysis and fitting of crystal electric field {Hamiltonians}},
