### 2025-11-29 a37c702
Added draft 'State of the Art' section
