### 2025-11-29 7207dbf
Update paper.md
