### 2025-11-29 bc08a5d
Update paper.md
