### 2026-02-20 #9176 @KedoKudo
> @KedoKudo I see you have marked all your checklist items as marked, but this issue opened by you on the software repo is still open https://github.com/AlbanBloorRiley/INS_IEP/issues/9. Can you please confirm if you believe that the submission is ready for acceptance?

They can workout the remaining issues at a later date. They are not blocking.
