### 2026-02-23 #9176 @mbarzegary
@AlbanBloorRiley given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with some minor edits (and check the changes)
* Work on the author's points of the final checklist I created ☝️
