### 2026-02-27 #9176 @AlbanBloorRiley
Version number is v1.0
zenodo DOI: 10.5281/zenodo.18805916
I have checked the authors and licence match
