### 2026-03-02 #9176 @mbarzegary
@AlbanBloorRiley The paper's bib file contains two bibliography entries that resolve to the same ID `baker_varying_2011`.  This prevents the acceptance script from proceeding. Please fix this issue. Also check that there is no other duplicate.
