+++ Paper/paper.bib
@@ -4,7 +4,6 @@
 @STRING{sisc="SIAM J. Sci. Comp."} 
 @STRING{siopt="SIAM J. Optim."}
 @STRING{siap="SIAM J. Appl. Math."} 
-
 @String{actanum="Acta Numer."}
 @String{actamath="Acta Math."}
 @String{invention="Invent. Math."}
@@ -258,21 +257,6 @@
 	pages = {141--143},
   }
 
-@article{baker_varying_2011,
-	title = {Varying spin state composition by the choice of capping ligand in a family of molecular chains: detailed analysis of magnetic properties of chromium(iii) horseshoes},
-	volume = {40},
-	issn = {1477-9226, 1477-9234},
-	shorttitle = {Varying spin state composition by the choice of capping ligand in a family of molecular chains},
-	url = {http://xlink.rsc.org/?DOI=c0dt01243b},
-	doi = {10.1039/c0dt01243b},
-	language = {en},
-	number = {12},
-	journal = {Dalton Transactions},
-	author = {Baker, Michael L. and Bianchi, Alberto and Carretta, Stefano and Collison, David and Docherty, Rebecca J. and Mclnnes, Eric J. L. and McRobbie, Andrew and Muryn, Christopher A. and Mutka, Hannu and Piligkos, Stergios and Rancan, Marzio and Santini, Paolo and Timco, Grigore A. and Tregenna-Piggott, Philip L. W. and Tuna, Floriana and Güdel, Hans U. and Winpenny, Richard E. P.},
-	year = {2011},
-	pages = {2725},
-	file = {Baker et al. - 2011 - Varying spin state composition by the choice of ca.pdf:/Users/user/Zotero/storage/UUQ6VSDW/Baker et al. - 2011 - Varying spin state composition by the choice of ca.pdf:application/pdf},
-}
 
 
 @incollection{baker_spectroscopy_2014,
@@ -293,7 +277,6 @@
 }
 
 
-
 @book{squires_introduction_2012,
 	edition = {3},
 	title = {Introduction to the {Theory} of {Thermal} {Neutron} {Scattering}},
@@ -308,7 +291,6 @@
 	year = {2012},
 	doi = {10.1017/CBO9781139107808},
 }
-
 
 @book{launay_electrons_2014,
 	address = {Oxford, United Kingdom},
@@ -329,7 +311,6 @@
 	author = {Baker, Michael L.},
 	year = {2022},
 }
-
 
 
 
