### 2026-03-15 #9176 @Kevin-Mattheus-Moerman
@AlbanBloorRiley As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

👉 Your project appears to lack detailed contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

👉 Please add the country information to your affiliations, and please do not use country acronyms like UK but spell out the name in full.
- [x] Copyedit pass/checked for typos

### Typos in the README
- [x]  `downladed` -> `downloaded`
- [x] `respository` -> `repository`
- [x] `instalation` -> `installation`
