### 2026-03-19 ef210fb
Create CONTRIBUTING.md
